[WIP] Introduce global read-only array

+++ src/main/scala/net/akouryy/synthescala/dsl/DSL.scala
@@ -1,5 +1,5 @@
 package net.akouryy.synthescala
 package dsl
 
-type U[T <: Int] = Long
-type S[T <: Int] = Long
+type U[T <: Int] = Int
+type S[T <: Int] = Int
